Updated GetHitRayLocation Method
Added some logic in this method to calculate the screen size using a engine method call GetViewportSize() Which out puts the size of which ever viewport the game is playing in.

It uses the output parameters which we created our own variables to use as place holders for the information it would output!